+++ core/src/stored/block.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2001-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2024 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -800,7 +800,10 @@ bool DeviceControlRecord::WriteBlockToDev()
   dcr->VolMediaId = dev->VolCatInfo.VolMediaId;
   if (dcr->VolFirstIndex == 0 && block->FirstIndex > 0) {
     dcr->WroteVol = true;
-    ASSERT(dcr->DirCreateJobmediaRecord(true));
+    if (!dcr->DirCreateJobmediaRecord(true)) {
+      Jmsg(dcr->jcr, M_ERROR, 0,
+           "Was not able to create dummy job media record.\n");
+    }
     dcr->VolFirstIndex = block->FirstIndex;
     uint64_t addr = dev->file_addr;